reimplement closure detection, drop map closures from the package
the detector lives in rko_slam/closures now and the package drops map
closures and hbst. opencv is a direct dependency, for the orb features and
the density image.

detection is as before: a bird's eye density image of the sub-map, orb
descriptors on it, ransac over the matches. what changes is how the map is
levelled before it is projected. with an imu the sub-map's measured up sets
the plane and the ground under it only fixes the height; without one the
ground fit does both, as it always did.

refinement moves to closures/refinement with the point to plane icp and the
voxel overlap gate unchanged. fill_sub_map needs ten points in a voxel for a
covariance now instead of three, which is what map closures itself used, so
the overlap score is computed over a sparser set and the 0.4 gate is a little
stricter than before.

part 2 of making this package releasable on rosdistro

+++ rko_slam/ros/base_node.cpp
@@ -188,8 +188,8 @@ BaseNode::BaseNode(const std::string& name, const rclcpp::NodeOptions& options) 
       node->declare_parameter<double>("density_threshold", slam_config.closure_detector.density_threshold));
   slam_config.closure_detector.hamming_distance_threshold = static_cast<int>(node->declare_parameter<std::int64_t>(
       "hamming_distance_threshold", slam_config.closure_detector.hamming_distance_threshold));
-  slam_config.closure_detector.inliers_threshold = static_cast<int>(
-      node->declare_parameter<std::int64_t>("inliers_threshold", slam_config.closure_detector.inliers_threshold));
+  slam_config.closure_detector.inliers_threshold = static_cast<std::size_t>(node->declare_parameter<std::int64_t>(
+      "inliers_threshold", static_cast<std::int64_t>(slam_config.closure_detector.inliers_threshold)));
   slam_config.closure_detector.no_of_sub_maps_to_skip = static_cast<int>(node->declare_parameter<std::int64_t>(
       "no_of_sub_maps_to_skip", slam_config.closure_detector.no_of_sub_maps_to_skip));
